Move TIME_WINDOW_DURATIONS to lib.scheduler_clients.models
sacct_base.py was the only one of 26 modules under slurm/cli_commands/
and pbs/cli_commands/ importing anything from a *_client module (it
pulled TIME_WINDOW_DURATIONS from slurm_base_client.py); every sibling
command module reaches only lib.exceptions,
lib.scheduler_clients.models, lib.ssh_clients.ssh_client, or another
command. That pointed the dependency the wrong way. JobsTimeWindow
already lives in lib.scheduler_clients.models, so the duration table
now sits right next to it, and both sacct_base.py and
slurm_rest_client.py import it from there instead of from a client
module. No behavior change; no import cycle existed before or after
(verified via a clean interpreter import of the whole package).

Also folds in the LAST_3_DAYS/LAST_7_DAYS "3days"/"7days" ->
"3d"/"7d" rename (for consistency with 1h/8h/24h) that was made
directly to models.py and partially to CHANGELOG.md alongside this
change; propagated the remaining "7days" references in
compute_slurm_api_test.py and CHANGELOG.md's breaking-change note to
match.
