+++ src/lib/scheduler_clients/models.py
@@ -5,7 +5,7 @@
 
 # models
 from enum import Enum
-from typing import List, Optional, Dict
+from typing import List, Optional, Dict, Tuple
 from lib.models import CamelModel
 
 from pydantic import Field, AliasChoices
@@ -15,8 +15,22 @@ class JobsTimeWindow(str, Enum):
     LAST_HOUR = "1h"
     LAST_8_HOURS = "8h"
     LAST_24_HOURS = "24h"
-    LAST_3_DAYS = "3days"
-    LAST_7_DAYS = "7days"
+    LAST_3_DAYS = "3d"
+    LAST_7_DAYS = "7d"
+
+
+# Canonical (amount, unit) duration for each historical time window. `unit`
+# doubles as both a `datetime.timedelta` keyword argument (used by the Slurm
+# REST client to compute an absolute start time) and a sacct/squeue
+# relative-time suffix (used by the Slurm CLI client to build
+# `--starttime=now-<amount><unit>`).
+TIME_WINDOW_DURATIONS: Dict[JobsTimeWindow, Tuple[int, str]] = {
+    JobsTimeWindow.LAST_HOUR: (1, "hours"),
+    JobsTimeWindow.LAST_8_HOURS: (8, "hours"),
+    JobsTimeWindow.LAST_24_HOURS: (24, "hours"),
+    JobsTimeWindow.LAST_3_DAYS: (3, "days"),
+    JobsTimeWindow.LAST_7_DAYS: (7, "days"),
+}
 
 
 class NodeState(str, Enum):
